Update Last Updated date and version

+++ build.ps1
@@ -7,13 +7,15 @@ param (
     [switch]$SkipReadMe,
     [Parameter(Mandatory, ParameterSetName="SkipPublishModule", HelpMessage="Skip publishing module to PSGallery")]
     [Parameter(ParameterSetName="SkipReadMe", HelpMessage="Skip publishing module to PSGallery")]
-    [switch]$SkipPublishModule
+    [switch]$SkipPublishModule,
+    [Parameter(ParameterSetName="SkipPublishModule", HelpMessage="Keep output")]
+    [switch]$KeepOutput
 )
 
 $scriptPath = split-path -parent $MyInvocation.MyCommand.Definition
 
 # Update Module Manifest
-$ModuleManifestPath = "$scriptPath\Intune.Graph.psd1"
+$ModuleManifestPath = "$scriptPath\Source\Intune.Graph.psd1"
 $ModuleManifest = Import-PowerShellDataFile $ModuleManifestPath
 $ModuleVersion = $ModuleManifest.ModuleVersion.Split(".")
 $UpdatedModuleVersion = "{0}.{1}.{2}" -f $ModuleVersion[0],$ModuleVersion[1],([int]$ModuleVersion[2]+1)
@@ -29,7 +31,7 @@ Update-ModuleManifest -Path $ModuleManifestPath -ReleaseNotes ""
 if($false -eq $SkipReadMe)
 {
     # Update README.md
-    $ReadmePath = "$scriptPath\..\README.md"
+    $ReadmePath = "$scriptPath\README.md"
     $Readme = Get-Content $ReadmePath
     $Readme[0] = "# Intune.Graph"
     $Readme[1] = "Last Updated: $((Get-Date).ToString('MM/dd/yyyy')) <br/>"
@@ -42,10 +44,7 @@ if($false -eq $SkipReadMe)
 # Build module
 $BuildPath = "$scriptPath\bin\Intune.Graph\$UpdatedModuleVersion"
 New-Item -ItemType Directory -Force -Path $BuildPath | Out-Null
-Copy-Item -Path "$scriptPath\Intune.Graph.psd1" -Destination $BuildPath
-Copy-Item -Path "$scriptPath\Intune.Graph.psm1" -Destination $BuildPath
-Get-ChildItem -Path "$scriptPath\" -Recurse -Directory | Copy-Item -Destination $BuildPath
-Get-ChildItem -Path "$scriptPath\" -Recurse -Include *.ps1 | Copy-Item -Destination $BuildPath 
+Copy-Item -Path "$scriptPath\Source\*" -Recurse -Destination $BuildPath
 
 if($false -eq $SkipPublishModule)
 {
@@ -63,5 +62,8 @@ if($false -eq $SkipPublishModule)
     Publish-Module @PublishParameters
 }
 
-# Remove build directory
-Remove-Item -Path "$scriptpath\bin" -Recurse -Force
+if($KeepOutput -eq $false)
+{
+    # Remove build directory
+    Remove-Item -Path "$scriptpath\bin" -Recurse -Force
+}
